+++ src/files/setup/setuplee.c
@@ -108,7 +108,7 @@
 #define AILIST_SET_UNCLOAKED                    0x0428
 
 s32 intro[];
-u8 props[];
+u32 props[];
 struct path paths[];
 struct ailist ailists[];
 
@@ -123,7 +123,7 @@ struct stagesetup setup = {
 	NULL,
 };
 
-u8 props[] = {
+u32 props[] = {
 	briefing(0, 0x2601)
 	briefing(1, 0x2600)
 	briefing(2, 0x2602)
